requirements.txt
pandas_datareader==0.10.0
pandas==2.1.4
numpy==1.26.3
tabulate==0.9.0
requests==2.31.0
matplotlib==3.3.4
python-dotenv==1.0.0
